+++ src/cpp_neural_lib/src/DenseLayer.cpp
@@ -101,6 +101,6 @@ void DenseLayer<Scalar>::zeroBias()
   }
 }
 
-template class DenseLayer<float>;
+template class DenseLayer<double>;
 template class DenseLayer<ADF>;
 template class DenseLayer<ADAD>;
