+++ src/cpp_neural_lib/src/types/Scalars.h
@@ -5,7 +5,7 @@
 
 //auto-derivative types. Pure psychosis.
 
-typedef CppAD::cg::CG<float> CGF;
+typedef CppAD::cg::CG<double> CGF;
 typedef CppAD::AD<CGF> ADCF;
 typedef CppAD::AD<ADCF> ADAD;
-typedef CppAD::AD<float> ADF;
+typedef CppAD::AD<double> ADF;
